src/main/java/com/microservice/user/security/CustomUserDetailsService.java
package com.microservice.user.security;

import com.microservice.user.user.domain.ports.UserPorts;
import com.microservice.user.user.domain.entities.User;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.Set;
import java.util.stream.Collectors;

@Service
public class CustomUserDetailsService implements UserDetailsService {

    private final UserPorts userPorts;

    public CustomUserDetailsService(UserPorts userPorts) {
        this.userPorts = userPorts;
    }

    @Override
    public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
          User user = userPorts.findByEmail(userEmail)
                 .orElseThrow(() ->
                         new UsernameNotFoundException("User not found with email: " + userEmail));

        Set<GrantedAuthority> authorities = user
                .getRoles()
                .stream()
                .map((role) -> new SimpleGrantedAuthority(role.getName()))
                .collect(Collectors.toSet());

        return new org.springframework.security.core.userdetails.User(user.getEmail(),
                user.getPassword(),
                authorities);
    }
}
